+++ src/api_calls/mod.rs
@@ -3,9 +3,10 @@ pub mod mod_calls;
 use super::structures::common_structs::Pagination;
 use serde::{Deserialize, Serialize};
 
-// For some reason CurseForge returns API responses in a 'response' json structure with the _actual_ results in `body`
-// Sometimes there are paginations so uhh yeah that too
+// CurseForge returns API responses in a 'response' json structure with the actual results in `body` and `pagination`s too
 #[derive(Deserialize, Serialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+#[serde(deny_unknown_fields)]
 pub(crate) struct Response<T> {
     data: T,
     pagination: Option<Pagination>,