+++ src/request.rs
@@ -1,16 +1,17 @@
 use crate::{Furse, Result};
-use reqwest::{IntoUrl, Response};
+use reqwest::{IntoUrl, Response, Url};
 
-const API_URL_BASE: &str = "https://api.curseforge.com/v1";
-
-/// Perform a GET request to the API base url + `route` using `client`
-pub(crate) async fn request_rel(client: &Furse, route: String) -> Result<Response> {
-    request(client, format!("{}{}", API_URL_BASE, route)).await
+lazy_static::lazy_static! {
+    pub(crate) static ref API_URL_BASE: Url = Url::parse("https://api.curseforge.com/v1/").unwrap();
 }
 
-/// Perform a GET request to `url` using `client`
+/// Perform a GET request on `url` using the HTTPS client and user agent from `client`
 pub(crate) async fn request(client: &Furse, url: impl IntoUrl) -> Result<Response> {
-    let request = client.client.get(url).header("x-api-key", &client.api_key);
-
-    request.send().await?.error_for_status()
+    Ok(client
+        .client
+        .get(url)
+        .header("x-api-key", &client.api_key)
+        .send()
+        .await?
+        .error_for_status()?)
 }